+++ CHANGELOG.md
@@ -2,6 +2,7 @@ unreleased
 * fixes #30 - dummy classifier does not predict by majority #30
 * added dummy regression model
 * improved design-matrix feature . Breaking !  columns need to be refered know by "precise name" (string, symbol, keyword)
+* made model  :fastmath/ols frezzable by nippy
  
 
 0.11.1